+++ SW07_ActorExample/ActorExample.cpp
@@ -7,7 +7,8 @@ void ActorExample::Initialize(HINSTANCE hInstance, LPCWSTR title, UINT width, UI
 	mspBackground = std::make_unique<Actor>(this, L"Data/back1.png");
 	for (int i = 0; i < 100; i++)
 	{
-		mBugs.push_back(std::make_unique<Actor>(this, L"Data/01_ecco_1.png", i * 10, i % 5 * 60, 1.0f, 0.09f));
+		mBugs.push_back(std::make_unique<Actor>(this, L"Data/01_ecco_1.png",
+			static_cast<float>(i * 10), static_cast<float>(i % 5) * 60, 1.0f, 0.09f));
 	}
 }
